add dataset selector, algorithm selector, update cost scaling chart with widths, cost calc method change.

+++ Src/TspEvo/moroutegraph.h
@@ -17,8 +17,8 @@ namespace MORouteGraph {
      (expressed by their coordinates)
      from the given file name */
 
-  float distance (unsigned __from, unsigned __to) ;
-  float cost(unsigned __from, unsigned __to);
+  double distance (unsigned __from, unsigned __to) ;
+  double cost(unsigned __from, unsigned __to);
   std::pair <double, double> getCityCoords(unsigned city);
    unsigned size () ; // How many cities ?
 }
